+++ dto/authorization_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019-2020 Authlete, Inc.
+// Copyright (C) 2019-2021 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -38,6 +38,9 @@ type AuthorizationResponse struct {
 
 	//
 	Scopes []Scope `json:"scopes,omitempty"`
+
+	//
+	DynamicScopes []DynamicScope `json:"dynamicScopes,omitempty"`
 
 	//
 	UiLocales []string `json:"uiLocales,omitempty"`