+++ dto/device_verification_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019-2020 Authlete, Inc.
+// Copyright (C) 2019-2021 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -36,6 +36,9 @@ type DeviceVerificationResponse struct {
 
 	//
 	Scopes []Scope `json:"scopes,omitempty"`
+
+	//
+	DynamicScopes []DynamicScope `json:"dynamicScopes,omitempty"`
 
 	//
 	ClaimNames []string `json:"claimNames,omitempty"`